Python update, bro. Nothing more, nothing less
